# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sugar Daddy Platform API
  description: |
    Comprehensive API documentation for the Sugar Daddy/Sugar Baby platform.
    This API provides authentication, user management, profile management, 
    matching, messaging, and subscription services.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@sugardaddyplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sugardaddyplatform.com/v1
    description: Production server
  - url: https://staging-api.sugardaddyplatform.com/v1
    description: Staging server
  - url: http://localhost:3001/v1
    description: Development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username
        firstName:
          type: string
          minLength: 1
          maxLength: 50
        lastName:
          type: string
          minLength: 1
          maxLength: 50
        role:
          type: string
          enum: [user, premium, admin]
          default: user
        isVerified:
          type: boolean
          default: false
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        code:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
paths:
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve user information by user ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...